requirements.txt
confluent-kafka==2.5.3
contourpy==1.3.0
cycler==0.12.1
Faker==29.0.0
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
